fixed some bugs & trying around with sseq

+++ sound/sound.h
@@ -25,7 +25,7 @@ public:
 
         std::vector<uint8_t> buffer;
 
-        uint32_t playPosition = 0; // Offset of reading, used by updateBuffer function
+        size_t playPosition = 0; // Offset of reading, used by updateBuffer function
         // Function pointer to update buffer
         //void (*updateBuffer)(size_t index, int len);
     };
